requirements.txt
aiohttp
python-telegram-bot
openai
python-dotenv
sqlalchemy
pytest